Unit tests for 64 bit ids and bug fixes from tests.
Create class SparsePage that is used for testing purposes to write
pages using the older page encoding format. Added unit tests that
verify 64-bit node identifiers are allowed. Fix minor errors caught
by these tests. Ref T51939.

+++ hydra-store/src/main/java/com/addthis/hydra/store/kv/TreeEncodeType.java
@@ -10,7 +10,7 @@ import com.addthis.basis.util.Bytes;
 
 public enum TreeEncodeType {
     BIT32(32, Integer.MAX_VALUE, "SPARSE"),
-    BIT64(64, Integer.MAX_VALUE, "LONGIDS");
+    BIT64(64, Long.MAX_VALUE, "LONGIDS");
 
     private final int bits;
